Add Terraform module for creating VMs for Hadoop cluster
The Terraform module creates a VM for running an instance of IPA server
to provide a Kerberos key distribution center (KDC) that supports
constrained delegation. VMs for running a NamdeNode server and multiple
DataNode servers are also created.

An initial debug Concourse task is provided for testing the creation and
deletion of the infrastructure.

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>