Add Ansible play for provisioning multinode Hadoop cluster
The Ansible installs IPA server on a dedicated VM and then installs
Hadoop (using singlecluster archive) on NameNode and DataNode VMs. The
Hadoop nodes are also configured as IPA clients.

A Concourse task for consuming the metadata produced by the Terraform
Concourse resource and generating configuration for Ansible is included.

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>