Detect malformed GPDBWritable formatted tuple
If the GPDBWritable message is malformed, it can cause the formatter to
attempt a read outside of the valid boundaries of data_duf. This results
in a SIGSEGV that panics the segment and leads to a failover.

This commit adds some basic bounds checks before reading and reports the
error by logging at the ERROR level with an error code of
ERRCODE_DATA_EXCEPTION. If the PXF external table has been defined with
LOG ERRORS, the malformed tuple can be viewed via gp_read_error_log.