Change log level to FATAL
Instead of logging at the ERROR level, we can log at the FATAL level.
This will keep the backend around for a short amount of time so that
someone can capture a core dump with either gcore or gdp. The amount of
time to keep the process around is configured via gp_debug_linger.

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>